Correct CSV DCG in list test
The contents of a CSV column may not contain a quote character.  The old
DCG allowed a single column to consume the content from its neighboring
column on backtracking.

I also added an optimizing cut, just to be thorough.  Once a column has
matched the closing quote character, backtracking will not provide any
additional solutions.  The cut makes this clear.